% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{coverage}
\usage{
coverage(data, paramter, lower, upper, subset = NULL)
}
\arguments{
\item{data}{}

\item{paramter}{}

\item{lower}{}

\item{upper}{}
}
\description{
coverage
}
