% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_multiple.R
\name{coverage_multiple}
\alias{coverage_multiple}
\title{coverage_multiple}
\usage{
coverage_multiple(data, y, lower, upper, subset = NULL, group_to_check)
}
\description{
when there are multiple obervations from same group eg. two observations
with the same country and year
}
