% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_data.R
\name{error_data}
\alias{error_data}
\title{error_data}
\usage{
error_data(data, parameter, estimate, totalerror_sd)
}
\arguments{
\item{data}{}

\item{parameter}{}

\item{estimate}{}

\item{totalerror_sd}{}
}
\description{
error_data
}
