% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals}
\alias{residuals}
\title{residuals}
\usage{
residuals(data, y, yhat, total_standard_error, subset = NULL)
}
\arguments{
\item{data}{\emph{\sQuote{Tibble}} formated as our example package data \code{\link[rvalidate:rv_data]{rvalidate:rv_data}}}

\item{y}{\emph{\sQuote{Character}} columm name of y in your data}

\item{yhat}{\emph{\sQuote{Character}} column name of yhat in your data}

\item{total_standard_error}{\emph{\sQuote{Character}} column name of total standard error in your data}
}
\value{
\emph{\sQuote{Tibble}} as a table with errors
}
\description{
\describe{
   \item{error (denoted \eqn{\epsilon}) = \eqn{y - \hat{y}}}
   \item{standard error = \eqn{\epsilon / sd(\epsilon)}}
   \item{adjusted error = \eqn{\epsilon / (1 / total standard error )}}
}
}
