% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_coverage.R
\name{rv_coverage}
\alias{rv_coverage}
\title{rv_coverage}
\usage{
rv_coverage(data, parameter, lower, upper, subset = NULL)
}
\arguments{
\item{data}{\emph{\sQuote{Tibble}} formated as our example package data \code{\link[rvalidate:rv_data]{rvalidate:rv_data}}}

\item{parameter}{\emph{\sQuote{Character}} columm name of parameter in your data}

\item{lower}{\emph{\sQuote{Character}} columm name of lower bound in your data}

\item{upper}{\emph{\sQuote{Character}} columm name of upper bound in your data}
}
\value{
\emph{\sQuote{Tibble}} as a table with coverage
}
\description{
rv_coverage
}
