% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_coverage_multiple.R
\name{rv_coverage_multiple}
\alias{rv_coverage_multiple}
\title{rv_coverage_multiple}
\usage{
rv_coverage_multiple(data, parameter, lower, upper, subset = NULL,
  group_to_check)
}
\description{
when there are multiple obervations from same group eg. two observations
with the same country and year
}
