% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rv_error.R
\name{rv_error}
\alias{rv_error}
\title{rv_error}
\usage{
rv_error(data, parameter, estimate, total_standard_error, subset = NULL)
}
\arguments{
\item{data}{\emph{\sQuote{Tibble}} formated as our example package data \code{\link[rvalidate:simdata]{rvalidate:simdata}}}

\item{parameter}{\emph{\sQuote{Character}} columm name of parameter in your data}

\item{estimate}{\emph{\sQuote{Character}} column name of estimate in your data}

\item{total_standard_error}{\emph{\sQuote{Character}} column name of total standard error in your data}
}
\value{
\emph{\sQuote{Tibble}} as a table with errors
}
\description{
\describe{
   \item{error (denoted \eqn{\epsilon}) = \eqn{y - \hat{y}}}
   \item{standard error = \eqn{\epsilon / sd(\epsilon)}}
   \item{adjusted error = \eqn{\epsilon / (1 / total standard error )}}
}
}
