% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{coverage}
\usage{
coverage(data, y, lower, upper, subset = NULL)
}
\arguments{
\item{data}{\emph{\sQuote{Tibble}} formated as our example package data \code{\link[rvalidate:data]{rvalidate:data}}}

\item{y}{\emph{\sQuote{Character}} columm name of y in your data}

\item{lower}{\emph{\sQuote{Character}} columm name of lower bound in your data}

\item{upper}{\emph{\sQuote{Character}} columm name of upper bound in your data}
}
\value{
\emph{\sQuote{Tibble}} as a table with coverage
}
\description{
coverage
}
